+++ android/assets/sprite_frenemy_copy.png
